refactor(hooks): layout에서만 사용되는 hooks을 공용 디렉토리에서 layout 하위로 이동
