+++ src/layout/Side/SideArea/SideArea.tsx
@@ -4,11 +4,11 @@ import { noop, clamp } from 'lodash-es'
 
 /* Internal dependencies */
 import useEventHandler from '../../../hooks/useEventHandler'
-import useLayoutState from '../../../hooks/useLayoutState'
-import useResizingHandlers from '../../../hooks/useResizingHandlers'
+import useLayoutState from '../../hooks/useLayoutState'
+import useResizingHandlers from '../../hooks/useResizingHandlers'
 import { document } from '../../../utils/domUtils'
 import { CONTENT_MIN_WIDTH, SIDE_MAX_WIDTH, SIDE_MIN_WIDTH } from '../../../constants/LayoutSizes'
-import useLayoutDispatch from '../../../hooks/useLayoutDispatch'
+import useLayoutDispatch from '../../hooks/useLayoutDispatch'
 import LayoutActions from '../../redux/LayoutActions'
 import { Resizer, ScrollWrapper, SideAreaWrapper } from './SideArea.styled'
 import SideAreaProps from './SideArea.types'
