+++ src/layout/hooks/useHeader.ts
@@ -1,7 +1,7 @@
 /* External dependencies */
 import { useEffect } from 'react'
-import LayoutHeaderType from '../types/LayoutHeaderType'
-import { LayoutActions } from '../layout/redux'
+import LayoutHeaderType from '../../types/LayoutHeaderType'
+import { LayoutActions } from '../redux'
 
 /* Internal dependencies */
 import useLayoutDispatch from './useLayoutDispatch'
