+++ src/layout/hooks/useSideWidth.ts
@@ -2,8 +2,8 @@
 import { useEffect } from 'react'
 
 /* Internal dependencies */
-import { SIDE_FALLBACK_WIDTH } from '../constants/LayoutSizes'
-import LayoutActions from '../layout/redux/LayoutActions'
+import { SIDE_FALLBACK_WIDTH } from '../../constants/LayoutSizes'
+import LayoutActions from '../redux/LayoutActions'
 import useLayoutDispatch from './useLayoutDispatch'
 
 export default function useSideWidth(width: number = SIDE_FALLBACK_WIDTH) {
