+++ src/layout/stories/LayoutPlayground/ContentHeader.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 /* Internal dependencies */
 import { styled } from '../../../foundation'
 import { Text } from '../../../components/Text'
-import useHeader from '../../../hooks/useHeader'
+import useHeader from '../../hooks/useHeader'
 import LayoutHeaderType from '../../../types/LayoutHeaderType'
 
 const Div = styled.div`
